Add support for Windows Runtime
This commit does not make any changes to the NuGet package.

The Options.WinRT project file contains only links to the existing
Options code. Support for F# was removed, as it does not exist in
WinRT. Similarly, support for serialization in NoneException was
removed.

This commit also updates the solution file to 2012.